requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
numpy==1.26.4
protobuf==4.25.3

SQLAlchemy~=2.0.27
pandas~=2.2.1
asyncpg~=0.29.0